splitup ephemeral_processing_with_export
