+++ src/actinia_core/rest/raster_export.py
@@ -31,7 +31,8 @@ from flask import jsonify, make_response
 import pickle
 from flask_restful_swagger_2 import swagger
 from actinia_core.rest.base.resource_base import ResourceBase
-from .ephemeral_processing_with_export import EphemeralProcessingWithExport
+from actinia_core.processing.actinia_processing.ephemeral.ephemeral_processing_with_export \
+     import EphemeralProcessingWithExport
 from actinia_core.core.common.redis_interface import enqueue_job
 from actinia_core.models.response_models import \
     ProcessingResponseModel, ProcessingErrorResponseModel
